requirements.txt
aiohttp
pydantic
loguru
python-Levenshtein